Refactored file structure and naming convention to match python recommended
